NEW AdminController as superclass for /admin/* routed controllers
