+++ tests/php/AdminControllerTest/DualPermissionAdminController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace SilverStripe\Admin\Tests\LeftAndMainTest;
+
+use SilverStripe\Admin\AdminController;
+use SilverStripe\Dev\TestOnly;
+
+class DualPermissionAdminController extends AdminController implements TestOnly
+{
+    private static string|array $required_permission_codes = [
+        'PERMISSION1',
+        'PERMISSION2',
+    ];
+}
